Удаление, добавление, поиск чего-либо кроме тестов

+++ Data/Interfaces/IRepository.cs
@@ -7,10 +7,10 @@ namespace Data.Interfaces
         where T: class
     {
         IEnumerable<T> GetListEntity(); // получение всех объектов
-        T GetEntity(int id);            // получение одного объекта по id
+        T GetEntity(long id);           // получение одного объекта по id
         void Create(T item);            // создание объекта
         void Update(T item);            // обновление объекта
-        void Delete(int id);            // удаление объекта по id
+        void Delete(long id);           // удаление объекта по id
         void Save();                    // сохранение изменений
     }
 }
